loading saving works now

+++ src/WorldStorage/NBTChunkSerializer.cpp
@@ -380,6 +380,7 @@ public:
 	{
 		mWriter.BeginCompound("");
 			AddBasicTileEntity(a_Entity,"Banner");
+
 			if (a_Entity->HasPatterns())
 			{
 				mWriter.BeginList("Patterns", TAG_Compound);
@@ -388,7 +389,7 @@ public:
 						mWriter.BeginCompound(std::to_string(i));
 						auto Pattern = a_Entity->GetPattern(i);
 						mWriter.AddString("Pattern", cBannerEntity::GetPatternTag(Pattern->m_Pattern));
-						mWriter.AddByte("Color", Pattern->m_Color);
+						mWriter.AddByte("Color", static_cast<UInt8>(Pattern->m_Color));
 						mWriter.EndCompound();
 					}
 				mWriter.EndList();
